consolidate workshop loading into custom hook
